# src/main/java/cjay/ai/cjreminder/dto/ReminderListRequest.java
package cjay.ai.cjreminder.dto;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;

@Getter
@NoArgsConstructor
@AllArgsConstructor
public class ReminderListRequest {

    private String name;
    private String color;
}